requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
httpx>=0.27
pydantic>=2.8
python-dotenv>=1.0
pytest>=8.2
